pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "markdown_webscraper_mcp"
version = "0.1.0"
description = "FastMCP server that uses markdown-webscraper to scrape websites and convert content to markdown."
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [{name = "Brosnan Yuen"}]
dependencies = [
  "fastmcp>=3.2.4",
  "markdown-webscraper>=0.1.4"
]

[project.optional-dependencies]
test = [
  "pytest>=9.0.0",
  "pytest-asyncio>=1.3.0"
]

[project.scripts]
markdown-webscraper-mcp = "markdown_webscraper_mcp.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
  "integration: integration tests requiring network and MCP server startup"
]
